feat: optimize partial fill (#56)
* Implementation of weightedPartialFillRate

* fix tests

* fix old tests

* Pass some test

* before refactor

* Fix up

* Issue w lowerCheckpoint

* pre-commit

* fix conflicts

* simplify exitPartiallyFilledBid

* Fix lastFullyFilledCheckpoint hint validation

* pre-commit

* Fix prev and next pointers in CheckpointStorage

* Add fuzz

* Simplify math

* Update tests all passing and pre-commit

* update pre-commit

* update readme

* feedback

* use `_getFinalCheckpoint` instead of `_unsafeCheckpoint`

* feedback

* rename checkpoint hints

* remove via-ir

+++ src/interfaces/ICheckpointStorage.sol
@@ -15,5 +15,5 @@ interface ICheckpointStorage {
     function currencyRaised() external view returns (uint128);
 
     /// @notice Get the number of the last checkpointed block
-    function lastCheckpointedBlock() external view returns (uint256);
+    function lastCheckpointedBlock() external view returns (uint64);
 }
